microservice/requirements.txt
async-timeout==4.0.3
bleach==6.0.0
blinker==1.6.2
cachelib==0.9.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Deprecated==1.2.14
flake8==6.1.0
Flask==2.3.3
Flask-Caching==2.0.2
Flask-Cors==4.0.0
Flask-Limiter==3.4.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.5.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mccabe==0.7.0
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
prometheus-client==0.17.1
prometheus-flask-exporter==0.22.4
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.16.1
PyMySQL==1.1.0
python-decouple==3.8
redis==5.0.0
requests==2.31.0
rich==13.5.2
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
urllib3==2.0.4
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.15.0
WTForms==3.0.1
zipp==3.16.2
prometheus-flask-exporter==0.22.4
prometheus_client
